change confirm password page layout

+++ src/app/signup/components/confirm-email/confirm-email.component.ts
@@ -8,6 +8,7 @@ import { ConfirmEmailService } from '../../services/confirm-email/confirm-email.
 import { HttpErrorResponse } from '@angular/common/http';
 import { ResendEmailPasscodeResponse } from '../../pojo/resend-email-passcode-response';
 import { StorageService } from '../../../shared/storage/storage.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-confirm-email',
@@ -22,11 +23,13 @@ export class ConfirmEmailComponent {
     private storageService: StorageService
   ) {}
 
+  public isLoad: boolean = false;
+
   public confirmEmailPasscodeForm: any = new FormGroup({
     passcode: new FormControl('', [Validators.pattern("^[0-9]{6}$")])
   })
 
-  public email: string = this.storageService.getItem("email") as string || "xxx@xxx";
+  public email: string = this.storageService.getItem("signup-email") as string || "xxx@xxx";
 
   public confirmEmailPasscodeFormSubmit() {
     if (this.confirmEmailPasscodeForm.status === "VALID") {
@@ -35,41 +38,47 @@ export class ConfirmEmailComponent {
       observable.subscribe({
         next: (response: PasscodeResponse) => {
           if (response.isPasscodeExpired) {
-            alert("Passcode expired");
+            Swal.fire("Passcode has expired",'','warning')
           }
           else {
             if (response.isPasscodeMatch) {
-              alert("Passcode CORRECT");
-              this.storageService.removeItem("signup-email");
-              this.router.navigate(["/signup"])
+              Swal.fire({
+                text: 'Passcode is correct. \nSign up successfully',
+                icon: 'success'
+              }).then((result)=>{
+                this.storageService.removeItem("signup-email");
+                this.router.navigate(["/signin"])
+              })
             }
-            else 
-              alert("Passcode FAIL");
+            else {
+              Swal.fire("Passcode does not match",'','warning')
+            }
           }
         },
         error: (e: HttpErrorResponse) => {
-          console.log("HttpServletResponse: " + e.error.message + "\n" + "ResponseEntity: " + e.error);
+          Swal.fire("An unknown error has happen. Please try again",'','error')
         }
       })
-    }
-    else {
-      alert("Incorrect passcode format")
     }
   }
 
   public reSendPasscode() {
+    this.isLoad = true;
     const observable: Observable<ResendEmailPasscodeResponse> = this.confirmEmailService.reSendPasscode(this.email);
     observable.subscribe({
       next: (response: ResendEmailPasscodeResponse) => {
         if (response.createdNewPasscode) {
-          alert("Create new passcode OK");
+          this.isLoad = false;
+          Swal.fire("New Passcode has been sent to your email",'','success');
         }
         else {
-          alert("Can not create new passcode");
+          this.isLoad = false;
+          Swal.fire("Error create new passcode. Please try again",'','error');
         }
       },
       error: (e: HttpErrorResponse) => {
-        console.log("HttpServletResponse: " + e.error.message + "\n" + "ResponseEntity: " + e.error);
+        this.isLoad = false;
+        Swal.fire("Error create new passcode. Please try again",'','error');
       }
     })
   }
